requirements.txt
alpaca-py
pandas
numpy
psycopg2-binary